добавлены схемы sql таблиц

+++ src/main/kotlin/su/redume/models/Crypto.kt
@@ -6,6 +6,6 @@ import org.jetbrains.exposed.sql.stringParam
 object Crypto : Table() {
     val fromCurrecny = stringParam("fromCurrecny")
     val convCurrency = stringParam("convCurrency")
-    val rate = integer("rate")
+    val rate = float("rate")
     val date = long("date")
 }
